Trying to make s390x work

+++ deprecated/int96_s390x.go
@@ -320,8 +320,9 @@ func BytesToInt96(data []byte) []Int96 {
 	for i := 0; i < len(result); i++ {
 		start := i * 12
 		for j := 0; j <= 2; j++ {
-			value := binary.LittleEndian.Uint32(data[start : start+4]) // Big endian conversion
-			result[i][j] = uint32(value)
+			value := binary.LittleEndian.Uint32(data[start : start+4])
+			//			result[i][j] = uint32(value)
+			result[i][j] = value
 			start += 4
 		}
 	}
